+++ src/locale/OutputTexts.ts
@@ -135,7 +135,7 @@ type OutputTexts = {
     };
     /** The base form type */
     Form: NameAndDoc;
-    /** A rectangle shape, for Stage.frame */
+    /** A rectangle form */
     Rectangle: NameAndDoc & {
         /** Left of the rectangle */
         left: NameAndDoc;
@@ -148,7 +148,7 @@ type OutputTexts = {
         /** Depth of rectangle */
         z: NameAndDoc;
     };
-    /** A rectangle shape, for Stage.frame */
+    /** A circle form */
     Circle: NameAndDoc & {
         /** Radius of the circle */
         radius: NameAndDoc;
@@ -157,6 +157,19 @@ type OutputTexts = {
         /** Vertical center of the circle */
         y: NameAndDoc;
         /** Z coordinate the circle */
+        z: NameAndDoc;
+    };
+    /** A regular polygon form */
+    Polygon: NameAndDoc & {
+        /** Radius of the polygon */
+        radius: NameAndDoc;
+        /** Radius of the polygon */
+        sides: NameAndDoc;
+        /** Horizontal center of the polygon */
+        x: NameAndDoc;
+        /** Vertical center of the polygon */
+        y: NameAndDoc;
+        /** Z coordinate the polygon */
         z: NameAndDoc;
     };
     /** A pose, for use in overriding an output's defaults for entering, resting, moving, or existing states */